Fix: cerchi pavimento rimossi + arma davanti al player + warning C4244
1. RIMOSSI CERCHI CHIARI SUL PAVIMENTO (Maze.cpp)
   I cerchi chiari (lightSpot) che apparivano come refusi di lanterne
   sul pavimento del labirinto sono stati rimossi. Erano cerchi di
   luce casuale (14px, alpha 90) su ~50% delle celle, che creavano
   un effetto visivo confuso.

2. ARMA DAVANTI AL PLAYER (Player.cpp)
   L'arma equipaggiata ora appare DAVANTI al player nella direzione di
   movimento, non dietro a sinistra come prima.
   - ramo sprite PNG: usa lastDx/lastDy per posizionare l'arma
     * dx > 0 (destra): arma a px + 14
     * dx < 0 (sinistra): arma a px - 14
     * dy > 0 (giu): arma in basso
     * dy < 0 (su): arma in alto
   - ramo fallback procedurale: stessa logica
   L'arma segue anche il salto (weaponY -= jumpOffset).

3. WARNING C4244 RESIDUI FIXATI (Game.cpp)
   - sinf(... + x * 0.01f + layer): cast (float)x e (float)layer
   - WINDOW_WIDTH/2.f -> (float)WINDOW_WIDTH/2.f
   - Altri cast espliciti int->float

Compila senza errori.

+++ src/Game.cpp
@@ -3537,7 +3537,7 @@ void Game::drawMenu() {
                                           sf::Color(40, 50, 100, 40);
         float yBase = WINDOW_HEIGHT - 180.f + layer * 30.f;
         for (int x = 0; x < WINDOW_WIDTH; x += 16) {
-            float y = yBase + sinf(menuTime * 0.5f + x * 0.01f + layer) * 15.f;
+            float y = yBase + sinf(menuTime * 0.5f + (float)x * 0.01f + (float)layer) * 15.f;
             sf::CircleShape fog(40.f);
             fog.setFillColor(fogCol);
             fog.setPosition((float)x - 40.f, y - 40.f);
@@ -3600,29 +3600,29 @@ void Game::drawMenu() {
     // Rombo centrale
     sf::ConvexShape diamond; diamond.setPointCount(4);
     diamond.setFillColor(ornGold); diamond.setOutlineThickness(1.f); diamond.setOutlineColor(sf::Color(120, 80, 20));
-    diamond.setPoint(0, sf::Vector2f(WINDOW_WIDTH/2.f, ornY - 6.f));
-    diamond.setPoint(1, sf::Vector2f(WINDOW_WIDTH/2.f + 8.f, ornY));
-    diamond.setPoint(2, sf::Vector2f(WINDOW_WIDTH/2.f, ornY + 6.f));
-    diamond.setPoint(3, sf::Vector2f(WINDOW_WIDTH/2.f - 8.f, ornY));
+    diamond.setPoint(0, sf::Vector2f((float)WINDOW_WIDTH/2.f, ornY - 6.f));
+    diamond.setPoint(1, sf::Vector2f((float)WINDOW_WIDTH/2.f + 8.f, ornY));
+    diamond.setPoint(2, sf::Vector2f((float)WINDOW_WIDTH/2.f, ornY + 6.f));
+    diamond.setPoint(3, sf::Vector2f((float)WINDOW_WIDTH/2.f - 8.f, ornY));
     window.draw(diamond);
     // Linee laterali
     sf::RectangleShape ornLineL(sf::Vector2f(180.f, 2.f));
     ornLineL.setFillColor(ornGold);
-    ornLineL.setPosition(WINDOW_WIDTH/2.f - 200.f, ornY - 1.f);
+    ornLineL.setPosition((float)WINDOW_WIDTH/2.f - 200.f, ornY - 1.f);
     window.draw(ornLineL);
     sf::RectangleShape ornLineR(sf::Vector2f(180.f, 2.f));
     ornLineR.setFillColor(ornGold);
-    ornLineR.setPosition(WINDOW_WIDTH/2.f + 20.f, ornY - 1.f);
+    ornLineR.setPosition((float)WINDOW_WIDTH/2.f + 20.f, ornY - 1.f);
     window.draw(ornLineR);
     // Piccoli rombi alle estremita' delle linee
     for (int side = 0; side < 2; side++) {
         float dx = (side == 0) ? -1.f : 1.f;
         sf::ConvexShape dot; dot.setPointCount(4);
         dot.setFillColor(ornGold);
-        dot.setPoint(0, sf::Vector2f(WINDOW_WIDTH/2.f + dx * 200.f, ornY - 4.f));
-        dot.setPoint(1, sf::Vector2f(WINDOW_WIDTH/2.f + dx * 204.f, ornY));
-        dot.setPoint(2, sf::Vector2f(WINDOW_WIDTH/2.f + dx * 200.f, ornY + 4.f));
-        dot.setPoint(3, sf::Vector2f(WINDOW_WIDTH/2.f + dx * 196.f, ornY));
+        dot.setPoint(0, sf::Vector2f((float)WINDOW_WIDTH/2.f + dx * 200.f, ornY - 4.f));
+        dot.setPoint(1, sf::Vector2f((float)WINDOW_WIDTH/2.f + dx * 204.f, ornY));
+        dot.setPoint(2, sf::Vector2f((float)WINDOW_WIDTH/2.f + dx * 200.f, ornY + 4.f));
+        dot.setPoint(3, sf::Vector2f((float)WINDOW_WIDTH/2.f + dx * 196.f, ornY));
         window.draw(dot);
     }
